refactor: getPathPrefixes
